requirements.txt
Django==4.2.3
gunicorn==21.2.0
psycopg2-binary==2.9.6
python-dotenv==1.0.1
requests==2.32.3
beautifulsoup4>=4.9.0
openai>=1.0.0
